refactor: make solution id as long & prevent double ai critique

+++ src/main/java/me/gabcytn/srsly/Controller/AiController.java
@@ -15,7 +15,7 @@ public class AiController {
   private final AiService aiService;
 
   @GetMapping("/{solutionId}/ai")
-  public AiCritique index(@PathVariable int solutionId) {
+  public AiCritique index(@PathVariable long solutionId) {
     return aiService.critique(solutionId);
   }
 }
